+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Add recursive calculation of Pearson correlation coefficent.
 * Change H-SAF reading interface to use pygeobase consistently.
 * H07 reader now returns more variables.
+* Resampling interface now respects dtype of input data.
 
 
 # v0.3.6, 2015-12-10